socialcommerce/requirement.txt
autopep8==1.3.3
Django==2.0
djangorestframework==3.7.7
olefile==0.44
Pillow==4.3.0
pycodestyle==2.3.1
pytz==2017.3
